[#14] feat : 가계부 수정 및 조회 권한 체크 로직 추가

+++ src/main/kotlin/com/account/manage/accountmanage/accountbook/application/service/AccountInfoUpdateService.kt
@@ -5,6 +5,7 @@ import com.account.manage.accountmanage.accountbook.adapter.out.persistence.Acco
 import com.account.manage.accountmanage.accountbook.application.port.`in`.AccountInfoUpdateUseCase
 import com.account.manage.accountmanage.accountbook.model.*
 import com.account.manage.accountmanage.common.adpater.out.error.DataNotFoundException
+import com.account.manage.accountmanage.user.domain.User
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import java.time.LocalDateTime
@@ -13,16 +14,20 @@ import java.time.ZoneId
 @Service
 class AccountInfoUpdateService(
     private val accountBookRepository: AccountBookRepository,
+    private val permissionCheckService: PermissionCheckService,
 ) : AccountInfoUpdateUseCase {
 
     @Transactional
     override fun updateAccountInfo(
-        accountBookRequest: AccountBookRequest, accountBookId: Long
+        accountBookRequest: AccountBookRequest, accountBookId: Long, user: User
     ): AccountBookResponse {
         val accountBook = accountBookRepository.findAccountBookById(accountBookId)
             ?: throw DataNotFoundException(AccountBookErrorType.NOT_FOUND_ACCOUNT_BOOK)
 
+        permissionCheckService.checkAccountBookPermission(user, accountBook)
+
         val accountUpdateDto = accountBookRequest as AccountUpdateDto
+
         accountBook.updateAccountBookInfo(accountUpdateDto.toVo(accountUpdateDto))
 
         return AccountUpdateCompletionDto(
@@ -33,10 +38,13 @@ class AccountInfoUpdateService(
 
     @Transactional
     override fun deleteAccount(
-        accountBookId: Long
+        accountBookId: Long, user: User
     ): AccountBookResponse {
         val accountBook = accountBookRepository.findAccountBookById(accountBookId)
             ?: throw DataNotFoundException(AccountBookErrorType.NOT_FOUND_ACCOUNT_BOOK)
+
+        permissionCheckService.checkAccountBookPermission(user, accountBook)
+
         accountBook.deleteAccountInfo()
 
         return AccountDeleteCompletionDto(
